feat(Karma): introduce `KarmaWithAccessControl`
This commit introduces a new `Karma` extension that adds `AccessControl`
capabilities. The reason this is done so that there can be multiple
actors in the system with different privileges.

The main changes done here are:

- Introduce internal functions for most of the Karma specific logic
  (this is necessary to allow properly extending the contract's
  functionality via modifiers)
- Add a `KarmaWithAccessControl` contract that inherits `Karma` and
  overrides the necessary functions
- Introduce `OPERATOR_ROLE` next to the already provided
  `DEFAULT_ADMIN_ROLE`

Due to storage layout conflicts, the extension is implemented by
inheriting `Karma` instead of adding the `AccessControl` functionality
to the `Karma` contract directly.

Closes #207